Mark various `Unix` fns as nonportable and make others portable (#4309)
* Mark various `Unix` fns as nonportable and make others portable

`fork` was already previously documented as not supported once domains have been
spawned.

Various `Unix` functions read the environment and `putenv` makes those unsafe.
Instead of marking a lot of functions as nonportable, we decided it is better to
mark `putenv` as nonportable and to alert that it is unsafe when multiple
domains are being used.

A bunch of other functions are marked as nonportable:
- `getprotobyname`,
- `getprotobynumber`,
- `getservbyname`, and
- `getservbyport`.

To make `getaddrinfo` and `getnameinfo` portable, the thread-unsafe "emulated"
versions are only used when running on the main domain.

This also implements a number of `Unix` functions using newer reentrant
functions assuming they exist on the target.

* Add trivial smoke tests for changed `Unix` fns using `*_r` APIs

Previously these functions were not tested at all.

* Add `@ once` and `@ local` on HOFs where possible