flambda2: Improve performance of free_names_transitive (#4684)
Currently, when processing a name, we first compute the free names of
its type and then recurse into the free names we haven't seen yet.

If we have `a` with free names `b` and `c`, and `b` with free names `c`,
this causes us to iterate on the free names of `c` twice (the first time
where we will process them, and the second time in the computation of
an empty `to_traverse`).

This patch instead checks if we have already processed a name
immediately before processing it, ensuring we only compute the free
names of a given name once.

This brings down the compilation of CFML's `parser.ml` (see #4674)
to 1min14s on my machine, from over 10min (I killed it) before.