Implement small array indices (#4691)
Add `.S(i)`, `.s(i)`, `.:S(i)`, and `.:s(i)` as the syntax for using `int16#`s and `int8#`s as array and iarray indices. Documentation has been updated in `jane/doc/extensions/_03-unboxed-types/03-block-indices.md`. This feature is straightforward since the implementation mirrors that of other array indicies. Most of the changes are syntactic. The semantic changes are located in `translprim.ml`. There is also a small change to the typechecker to type small int indices (naturally located in `typecore.ml`).

---------

Co-authored-by: James Rayman <james@jamesrayman.com>