Untagged char literals (#4658)
Add untagged chars literals (of the form `#'a'`). They are enabled with the small numbers extension[^1]. Documentation has been updated in `jane/doc/extensions/_11-miscellaneous-extensions/small-numbers.md`. Char# literals are supported in patterns, though not yet in ranges (will be implemented in a separate PR).

Aside from the testsuite and documentation, the most interesting change is in the lexer. Char# literals are a new type of token, though most of the lexer code is similar to that of regular char literals. Most of the other changes mirror the handling of regular chars or unboxed int8s.

[^1]: The type `char#` is always enabled since disabling only the unboxed version of a type would complicate the typechecker.

---------

Co-authored-by: James Rayman <james@jamesrayman.com>