datalog: Add provenance information for debugging (#4818)
This adds support for displaying the reason a particular fact was
derived by the datalog engine. This information must be enabled by
setting `enable_provenance_for_debug` early during module loading time
(it is currently enabled by the reaper using the existing environment
variable `READPERDBG` if it contains the value `prov`): since most
datalog rules are compiled when loading the reaper module, it is
complicated to use a runtime flag without degrading the performance when
not computing provenance information.