Unify the domain locks used by the backup thread and by systhreads (#4769)
Previouly, systhreads running OCaml code needed to hold both the systhreads lock,
and the unrelated domain_lock. The domain_lock (but not the systhreads lock) was
also held by the backup thread when it was performing GC. Managing both locks is
tricky, slow, and causes problems with the switch_runtime_locking_scheme system
used for Python interop. Consequently, switch_runtime_locking_scheme was disabled
whenever multiple domains are in use.

This patch unifies the two locks, by adding lock/unlock hooks to domain.c which
are overriden by st_stubs.c from systhreads, so that there is only one lock that
manages access to the domain's state and GC.

There is a separate lock and condition variable that manages communication with the
backup thread. This lock is never held while doing GC: it is purely for waking the
backup thread, which will drop it and take the domain lock if it turns out there is
work to do. The new mechanism only wakes the backup thread when an interrupt occurs,
which should be much more efficient that the prior mechanism which wakes it on every
blocking section.

More detailed notes on changes:

  - `domain.c`: Refactor backup thread interaction and add new lock/unlock hooks.
    Since the implementation of these hooks uses Caml_state->id to determine which
    domain to lock, some code rearrangement was needed to ensure these fields are
    initialised early enough.

  - `signals.c`: Only communicate with the backup thread while actually holding the
    lock. (I think this might fix a bug - the order of events here was backwards)

  - `st_stubs.c`: The default locks are now initialised at startup, since they are
    needed very early in domain initialisation. Backup thread communication is moved
    into the blocking section hooks (mirroring the runtime proper, instead of being
    part of the st_masterlock in st_pthreads.h). `domain_lockmode` is gone, and
    `caml_thread_use_domains` is a no-op to be removed later.

  - Testsuite:

    * c-api/test_c_thread_has_lock_systhread: now works as upstream
    * lib-domain/spawn_outside_main_thread: now works as upstream
    * lib-domain/use_domains_outside_main_thread: deleted (tests an error that no longer occurs)
    * lib-systhreads/swapgil_then_domains.ml: deleted, replaced with domains version of main swapgil test

There will be a follow-up PR to extend the `struct caml_locking_scheme` with a
callback for custom backup thread interaction, but it's not yet done.