runtime: use a skiplist for natdynlink roots (#4949)
When dynlinking libraries, modules' GC roots are stored in a linked
list. While iteration over this list (during GC) is OK, the list is
checked for duplicates when registering a new module, which is slow.

This patch turns the linked list into a skiplist, which allows for
more efficient lookup (and similar iteration complexity). This data
structure is also used to track (generational) global roots.

There's one catch: where in the linked list based implementation,
the `roots_mutex` is only held while creating a copy of the list
head before iteration, we now hold the lock during iteration of
the whole skiplist. This is similar to the code in
`caml_scan_global_roots` which also holds the lock while iterating
over all global roots skiplists.