Untagged char range patterns (#4729)
Allow range patterns for untagged chars, like `#'a'..#'z'`. The parser
is updated to recognize hash constants in intervals, and typecore is
updated to expand `#'a'..#'z'` into `#'a' | #'b' | .. | #'z'`.

I originally also included `int8` and `int8#` ranges in this feature,
but the signedness of literals causes some odd results when ranges are
implemented the easy way. For example:
```
match 10s with
| 0s .. 0xF0s -> foo
| _ -> bar
```
evaluates to `bar` since `0xF0s` is `-16s`, thus `0s .. 0xF0s` = `0s ..
-16s` = `-16s .. 0s`, and `10s` is not in that range.

This PR also contains proper match exhaustion for small ints and int#:
Take for example:
```
Warning 8 [partial-match]: this pattern-matching is not exhaustive.
Here is an example of a case that is not matched:
#6m
```
Previously, the output would've been:
```
Warning 8 [partial-match]: this pattern-matching is not exhaustive.
Here is an example of a case that is not matched:
_
```

Also, a match like:
```ocaml
match x with
| -128s | -127s | .. | 127s -> 0
```
is now recognized to be exhaustive.