Initial slambda with trivial transl and eval (#4776)
* Initial slambda with trivial transl and eval

Introduces the slambda IR with a single constructor `SLquote` that
simply wraps lambda. Includes a static evaluator that simply unwraps
the quote. The transl phase now outputs slambda and the various driver
pipelines have been updated to include static evaluation.

The new `-dslambda` flag prints slambda code before it's evaluated to
lambda.

* Modernise names

Changed the `SLquote` constructor to `Quote` and the `slambda` type to `t`.
Introduced the `SL` alias in modules that use `Slambda` so that `SLquote` can be
written as `SL.Quote`.

* Add profiling

* Finish modernising names (fix build)

* Fix formatting

---------

Co-authored-by: Jack Rickard <jrickard@janestreet.com>