Add `make clean` and `make distclean` targets (#5035)
- add a `make clean` target to remove build outputs without reconfiguring
- add a `make distclean` target to fully clean and drop `configure`-generated files

This addresses the issue of `git`-based cleanup:
* `git`-based cleanup deletes personal editor config files that are not tracked by git
* If you make git track them (like `.vscode/`) then everyone is forced to use the same config
This `make clean` and `make distclean` are alternative cleanup commands that only delete a specific list of build generates files, and leave the rest alone, meaning `.vscode/` is no longer deleted even if it is not tracked. This allows https://github.com/oxcaml/oxcaml/pull/4186 to go through.