chamelon: allow stopping minimization with `Ctrl-C` (#5061)
We currently use `Sys.command`, which internally calls the `system` C
function, which in turn blocks (ignores) the SIGINT signals from hitting
`Ctrl-C` during the evaluation of the underlying compilation command.

This makes it hard to stop minimization; the underlying compilation
command does receive a SIGINT signal when the user hits `Ctrl-C`, but we
cannot robustly detect it from the result of `Sys.command`.

This patch replaces the use of `Sys.command` with calls to the `Unix`
module. The `Unix` module provides better information on the exit status
of the underlying command, but we don't even need to make use of it
since `Unix.create_process` (unlike `Sys.command`) does not use `system`
and hence does not mask SIGINT.

I did not find a way to be able to use the `Unix` module when chamelon
is built using the boot compiler without resorting to some dark `dune`
invocations (which we already use elsewhere for this purpose -- credit
to Mark Shinwell for both the trick of passing `unix.cmxa` to
`ocamlopt_flags` instead of adding `unix` to the `libraries`, and for
pointing me to the `context_name` variable).