Add profiling counters for inlining that does not decrease code size (#5116)
This patch adds a profiling counter for situations where we decide to
inline a function call due to speculative inlining, but said inlining
does not decrease the code size of the inlined function.

This is a situation that should not happen and is likely caused by a bug
(at least with configurations where the `small_function_size` and the
speculative `threshold` are identical, which they are for default
configurations).

The current ratios of speculative inlining decisions that do not
decrease the code size of inlined functions are:

 - 60/159 (~40%) on the stdlib
 - 438/2996 (~15%) on the compiler itself

These numbers will be reduced in follow-up PRs that fix bugs in
the computation of cost metrics.