Do not overestimate removed primitives due to canonicalization (#5118)
In `Simplify_named`, when simplifying a primitive, we consider that it 
has been removed unless we keep a let-binding with the primitive (in its
original form) as its right-hand side. This means that if the arguments
of the primitive have been canonicalized (typically, this is the case
when the primitive applies to a parameter of a function that gets
inlined), we incorrectly count it as having been removed.

This patch changes the equality check used in `Simplify_named` to take
into account canonicalization, so that a primitive whose arguments are
rewritten into a canonical form is not counted as having been removed.

This reduces the number of situations where speculative inlining does
not reduce the code size of the inlinined function. These situations
occur (ratio are number of successful speculative inlining that does not
reduce the code size / number of successful speculative inliing):

Without this change,

 - 60/159 times (~40%) on the stdlib
 - 438/2996 times (~15%) on the compiler itself

With this change,

 - 31/115 times (~28%) on the stdlib
 - 216/2640 times (~8%) on the compiler itself

Note that we prevented *more* speculative inlining than simply the cases
where there is no reduction in code size of the inlined function: on the
stdlib (resp. the compiler), we now prevent inlining of 159 - 115 = 44
(resp. 2996 - 2640 = 356) functions that were previously inlined, but
only 60 - 31 = 29 (resp. 438 - 216 = 222) did not reduce the code size
of the inlined function.

This is normal: fixing the computation of the removed operations will
also prevent situations where we *do* make real optimizations, but the
bogus removals pushed us below the threshold to actually inline
(consider for instance a function with code size 107 where inlining
truly removes 2 primitives and 10 primitives applied to arguments were
erroneously counted as being removed).