Fix computation of code size when converting switch to lookup table (#5120)
... or affine expressions.

The computation of the code size in `Simplify_switch_expr` for switches
that are recognized as "single arg to same destination" switches and
converted to a lookup table or is just plain wrong, as it tries to
compute a difference in code size rather than the code size of the
rebuilt expr.

This can make inlining significantly more aggressive than intended for
functions that contain such switches; in fact this PR includes tests
where these constructs lead to functions with negative code sizes!