Fix removed operations count of switches (#5144)
Switches in flambda2 are always performed on a naked immediate
scrutinee, and sometime encode a `%tag_immediate` primitive (confusingly
recorded as an "identity switch").

We currently implement this by unconditionally introducing a tagged
version of the scrutinee before simplifying the switch, so that if the
switch is actually an identity switch, we can find that primitive in the
CSE environment. This primitive is also used for "boolean not" switches,
affine switches, and switches that can be converted to a lookup table.

In other words, a switch:

```fexpr
switch x with
| 0 -> k e0
| 1 -> k e1
| 2 -> k e2
```

is always simplified as:

```fexpr
let tagged_scrutinee = %tag_immediate x in
switch x with
| 0 -> k e0
| 1 -> k e1
| 2 -> k e2
```

This means that whenever we simplify a switch that is not eliminated
(into a `%tag_immediate` or `%boolean_not`, or an array load or an
affine expression) we count the removal of the `%tag_immediate`
primitive as a removed operations, potentially causing speculative
inlining to trigger even though no optimisations were made.

This patch changes the logic for introducing the `%tag_immediate`
primitive to follow the same pattern as we use for `%boolean_not`
and affine expressions: instead of optimistically introducing it in the
downwards pass, we simply look it up in the CSE environment in the
upwards pass and introduce it if it does not exist.

Note that this means that in situations where we have a switch that can
be simplified into a `%tag_immediate` primitive, followed by an actual
`%tag_immediate` primitive (or another such switch) in the same
function, we will fail to CSE the `%tag_immediate` primitive. This
situation should be relatively rare, though.